requirements.txt
python-dotenv
supabase
requests
streamlit
pandas
rapidfuzz
spacy
